+++ test/Unit/RuleSet/Php55Test.php
@@ -19,6 +19,7 @@ use Ergebnis\PhpCsFixer\Config\Name;
 use Ergebnis\PhpCsFixer\Config\PhpVersion;
 use Ergebnis\PhpCsFixer\Config\Rules;
 use Ergebnis\PhpCsFixer\Config\RuleSet;
+use ErickSkrauch\PhpCsFixer;
 use PHPUnit\Framework;
 
 #[Framework\Attributes\CoversClass(RuleSet\Php55::class)]
@@ -35,7 +36,7 @@ final class Php55Test extends ExplicitRuleSetTestCase
 {
     protected function expectedCustomFixers(): Fixers
     {
-        return Fixers::empty();
+        return Fixers::fromFixers(new PhpCsFixer\Whitespace\LineBreakAfterStatementsFixer());
     }
 
     protected static function createRuleSet(): RuleSet
@@ -51,6 +52,7 @@ final class Php55Test extends ExplicitRuleSetTestCase
     protected function expectedRules(): Rules
     {
         return Rules::fromArray([
+            'ErickSkrauch/line_break_after_statements' => true,
             'align_multiline_comment' => [
                 'comment_type' => 'all_multiline',
             ],